refactor: reorganize file parsing
